requirements.txt
haystack-ai
python-dotenv
huggingface-hub
pypdf
sentence-transformers
accelerate
nltk